Added TakenDamageIndicator
Added TakenDamageIndicator, now when the character is hit, there will be a sign "-5" (the number is damage value and not hard-coded) at the end of the HealthBar. The indicator will gradually fade out in 2 secs. If, in this 2 secs, the character is hit again, the indicator will restart fade away "animation".
Added a destroy() method in UnitScript, which allows to do some necessary things before truly destroying the gameObject.