Added a Transition Screen
Added a transition screen, when switching control, there will be a transition screen pop up, and the game will proceed only after the player click on I'm Ready! button.
Note that since this is a ScreenSpace UI, if you want to what the game looks like in the Game view port, simply make the TransitionCanvas disabled. And remember to enable it when playing, or it will not behave properly.