Fixed the Bug In Team Manager
Fixed the bug in the Team Manager.
Before this update, after removing every unit from the selectionDict, all removed units will be tagged "Untagged".
Since we are going to get rid of the marquee selection, these code that tag removed units to "Untagged" will be removed in the future. I just comment them out to make the Team and Turn Manager work for now.